UIKIT-299: wip(ui,DataGrid): Добавлен пропс для пустой ячейки. Изменен формат компонента по style guid'у
