--- packages/ui/src/DataGrid/DataGridCell/types.ts
@@ -1,6 +1,0 @@
-import { DataGridColumns } from '../types';
-
-export type CellProps<Data> = {
-  row: Data;
-  cell: DataGridColumns<Data>;
-};
