+++ packages/ui/src/DataGrid/DataGridHead/index.ts
@@ -1,3 +1,1 @@
 export * from './DataGridHead';
-
-export * from './types';
