+++ packages/ui/src/DataGrid/DataGridHeadColumn/DataGridHeadColumn.tsx
@@ -1,9 +1,24 @@
-import { useMemo } from 'react';
+import { CSSProperties, useMemo } from 'react';
+import { TableCellProps } from '@mui/material';
+
 import { Typography } from '../../Typography/Typography';
 import { SortStates } from '../constants';
+import { DataGridRow, DataGridSort } from '../types';
 
 import { StyledTableCell, StyledTableSortLabel } from './styled';
-import { DataGridHeadColumnProps } from './types';
+
+export type DataGridHeadColumnProps<
+  Data = DataGridRow,
+  SortField extends keyof Data = keyof Data,
+> = {
+  onSort: (field: SortField, sortable?: boolean) => void;
+  sorting: DataGridSort<SortField>[];
+  label?: string;
+  sortable?: boolean;
+  align?: TableCellProps['align'];
+  field?: keyof Data;
+  width?: CSSProperties['width'];
+};
 
 export function DataGridHeadColumn<Data, SortField extends keyof Data>({
   onSort,
@@ -14,9 +29,15 @@ export function DataGridHeadColumn<Data, SortField extends keyof Data>({
   sorting,
   width,
 }: DataGridHeadColumnProps<Data, SortField>) {
-  const sortParams = useMemo(() => sorting.find(({ fieldId }) => field === fieldId), [sorting, field]);
-  const hideSortIcon = useMemo(() => !Boolean(sortParams),[sortParams]);
-  const sortDirection =  useMemo(() => sortParams ? sortParams.sort : SortStates.ASC, [sortParams]);
+  const sortParams = useMemo(
+    () => sorting.find(({ fieldId }) => field === fieldId),
+    [sorting, field],
+  );
+  const hideSortIcon = useMemo(() => !Boolean(sortParams), [sortParams]);
+  const sortDirection = useMemo(
+    () => (sortParams ? sortParams.sort : SortStates.ASC),
+    [sortParams],
+  );
 
   const handleSortClick = () => {
     if (field) {
