+++ packages/ui/src/DataGrid/DataGridHeadColumn/index.ts
@@ -1,3 +1,1 @@
 export * from './DataGridHeadColumn';
-
-export * from './types';
