+++ packages/ui/src/DataGrid/DataGridLoader/index.ts
@@ -1,3 +1,1 @@
 export * from './styled';
-
-export * from './types';
