--- packages/ui/src/DataGrid/DataGridLoader/types.ts
@@ -1,3 +1,0 @@
-export type DataGridLoaderProps = {
-  loading?: boolean;
-};
